test: address review feedback on integration suite
- config: add allowPublicKeyRetrieval to mysql configs (hardens caching_sha2
  auth over non-TLS across MySQL images/versions)
- helpers: extract runFile.js to remove triplicated connection-lifecycle
  boilerplate from the three test files
- featureSuite: add coverage for append({ unsafe: true }) literal interpolation
- docker-compose: drop obsolete `version` key
