+++ docker-compose.yml
@@ -1,4 +1,3 @@
-version: '3.9'
 services:
   postgres:
     image: postgres:17-alpine
