+++ integration/SQL.pg.integration.test.js
@@ -1,18 +1,6 @@
 'use strict'
 
-const { test, before, after } = require('node:test')
-const { withPg, createUsersTable } = require('./helpers/db')
-const runFeatureSuite = require('./shared/featureSuite')
+const { withPg } = require('./helpers/db')
+const runIntegrationFile = require('./helpers/runFile')
 
-const ctx = { db: null }
-
-before(async () => {
-  ctx.db = await withPg()
-  await createUsersTable(ctx.db)
-})
-
-after(async () => {
-  if (ctx.db) await ctx.db.end()
-})
-
-runFeatureSuite(test, () => ctx.db)
+runIntegrationFile(withPg)
